# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart LPR API",
        "description": "Open API for Smart LPR",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/history/{date_start}/{date_end}/{gate}/{page}/{limit}": {
            "get": {
                "tags": [
                    "History"
                ],
                "summary": "Get history of items",
                "description": "Retrieve a history of items based on date range, gate, pagination, and limit",
                "operationId": "history",
                "parameters": [
                    {
                        "name": "date_start",
                        "in": "path",
                        "description": "Start date for filtering the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2024-09-01 00:00:00"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "path",
                        "description": "End date for filtering the list",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "datetime",
                            "example": "2024-09-05 23:59:59"
                        }
                    },
                    {
                        "name": "gate",
                        "in": "path",
                        "description": "Gate number for filtering the list",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "66d9965ec312805ebe04d8bd"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/v1/webhook": {
            "get": {
                "tags": [
                    "webhook"
                ],
                "summary": "Get all webhooks",
                "operationId": "f13c7aa9ea14174035aa4f72459c0ecd",
                "responses": {
                    "200": {
                        "description": "List all webhooks"
                    }
                }
            },
            "post": {
                "tags": [
                    "webhook"
                ],
                "summary": "Retrieve the specified webhook using form data",
                "operationId": "97f3d31ca0266652bb539a2bab609c8f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "webhook": {
                                        "description": "ID of the history to retrieve",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retrieve the specified webhook"
                    },
                    "400": {
                        "description": "webhook not found"
                    }
                }
            }
        },
        "/api/v1/webhook/update": {
            "post": {
                "tags": [
                    "webhook"
                ],
                "summary": "Update the specified webhook using form data",
                "operationId": "2605b5212b508d657eda143299413f7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "webhook": {
                                        "description": "Webhook URL to update",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "id": {
                                        "description": "ID of the webhook to update",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook updated successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/webhook/{id}": {
            "delete": {
                "tags": [
                    "webhook"
                ],
                "summary": "Delete the specified webhook",
                "operationId": "fafcfff6182492535257d1e40cd2f5d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the webhook to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1a2b3c4d"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook deleted successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Webhook not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}